bug 1440745 - move alt-adi apps to separate module
Maintenance for FetchADIFromHiveCronApp is a nightmare right now. It
behooves us to keep it completely isolated from the rest of the codebase.
This moves things around to isolate it.
