+++ socorro/unittest/cron/jobs/test_fetch_adi_from_hive.py
@@ -386,28 +386,3 @@ class TestFetchADIFromHive(IntegrationTestBase):
         )
         count, = pgcursor.fetchone()
         assert count == 0
-
-
-class TestFAKEFetchADIFromHive(IntegrationTestBase):
-
-    def _setup_config_manager(self):
-        return super(TestFAKEFetchADIFromHive, self)._setup_config_manager(
-            jobs_string=(
-                'socorro.cron.jobs.fetch_adi_from_hive.FAKEFetchADIFromHiveCronApp|1d'
-            ),
-        )
-
-    def test_mocked_fetch(self):
-        config_manager = self._setup_config_manager()
-
-        with config_manager.context() as config:
-            tab = CronTabberApp(config)
-            tab.run_all()
-
-            information = self._load_structure()
-            assert information['fetch-adi-from-hive']
-            assert not information['fetch-adi-from-hive']['last_error']
-
-            config.logger.info.assert_called_with(
-                'Faking the fetching of ADI from Hive :)'
-            )
